backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
pymysql==1.1.1
pypdf==5.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic-settings==2.8.1
